Extract Downloader + error hierarchy, drop Python 3.9 support
- antenati_downloader.py: Downloader (the class formerly known as
  AntenatiDownloader), ProgressBar dataclass and DEFAULT_* constants
  live here. The orchestration logic now uses antenati_http and
  antenati_iiif exclusively, so the module is GUI-friendly (no argparse,
  click or tqdm dependencies).
- antenati_errors.py: AntenatiError base + ManifestError /
  WafChallengeError subclasses introduced ahead of the hardening PR,
  plus ThreadError relocated from antenati.py.
- antenati.py: collapses to the CLI surface (argparse + main()) plus
  backward-compatible re-exports. `AntenatiDownloader` is now a public
  alias for `Downloader`; the GUI and any external scripts keep working
  unchanged. `run_cli` becomes a module-level helper instead of a
  Downloader method (it depends on tqdm, which the Downloader shouldn't).

Python 3.10 baseline:
- README: bumped declared requirement.
- ruff target-version: py310; the UP007/UP045 ignores for Optional are
  removed, switching annotations to `X | None`.
- The legacy per-file-ignores and the `antenati` mypy override are gone
  now that antenati.py is rewritten. antenati_gui.py keeps its relaxed
  override pending the GUI worker refactor.

Side effect of mypy strictness on the new module: a latent None-charset
issue (silent fallback only on Antenati's well-behaved server) is now
handled explicitly with `or 'utf-8'` -- semantically equivalent to
`bytes.decode()`'s default.

Tests:
- One test updated (`test_run_cli_uses_tqdm_progress_bar`) to call the
  new module-level `antenati.run_cli`. All 45 offline tests pass.
